long long SUMOTime, refs #1728
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@18213 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/traci-server/TraCIServerAPI_Simulation.cpp
@@ -86,7 +86,7 @@ TraCIServerAPI_Simulation::processGet(TraCIServer& server, tcpip::Storage& input
     switch (variable) {
         case VAR_TIME_STEP:
             tempMsg.writeUnsignedByte(TYPE_INTEGER);
-            tempMsg.writeInt(MSNet::getInstance()->getCurrentTimeStep());
+            tempMsg.writeInt((int)MSNet::getInstance()->getCurrentTimeStep());
             break;
         case VAR_LOADED_VEHICLES_NUMBER:
             writeVehicleStateNumber(server, tempMsg, MSNet::VEHICLE_STATE_BUILT);
@@ -147,7 +147,7 @@ TraCIServerAPI_Simulation::processGet(TraCIServer& server, tcpip::Storage& input
             break;
         case VAR_DELTA_T:
             tempMsg.writeUnsignedByte(TYPE_INTEGER);
-            tempMsg.writeInt(DELTA_T);
+            tempMsg.writeInt((int)DELTA_T);
             break;
         case VAR_NET_BOUNDING_BOX: {
             tempMsg.writeUnsignedByte(TYPE_BOUNDINGBOX);
